Combine all destination actions to one file, remove debugging logs
