+++ src/reducers/destinationReducer.js
@@ -1,5 +1,4 @@
 import * as destinationActions from '../actions/destination'
-import * as actions from '../actions/actions'
 import initialState from './initialState'
 
 function destinationReducer(state = initialState, action) {
@@ -20,7 +19,7 @@ function destinationReducer(state = initialState, action) {
                 destinationSnow: 'N/A',
                 destinationWeatherIcon: 'N/A'
             })
-        case actions.SET_DESTINATION:
+        case destinationActions.SET_DESTINATION:
             console.log('hit set destination reducer')
             console.log(action.regionDestination, 'action region')
             console.log(action.cityDestination, 'action city')
@@ -29,19 +28,13 @@ function destinationReducer(state = initialState, action) {
                 cityDestination: action.cityDestination,
                 latLngDestination: action.latLngDestination
             })
-        // case actions.SET_DESTINATION_ROUTE:
-        //     console.log('hit setDestinationRoute reducer');
-        //     console.log(action.setDestinationRoute, 'directions');
-        //     return Object.assign({}, state, {
-        //         destinationRoute: action.destinationRoute
-        //     })
-        case actions.SET_RETURN_DESTINATION:
+        case destinationActions.SET_RETURN_DESTINATION:
             return Object.assign({}, state, {
                 cityReturnDestination: action.cityReturnDestination,
                 regionReturnDestination: action.regionReturnDestination,
                 latLngReturnDestination: action.latLngReturnDestination
             })
-        case actions.HAVE_LOCATIONS:
+        case destinationActions.HAVE_LOCATIONS:
             return Object.assign({}, state, {
                 haveLocations: action.haveLocations
             })
